Uploaded LTspice files completed by 741 group
The four LTspice files upload cover the circuits that have been simulated by the 741 group so far during sprints 1-3. These files include simulations of a basic RLC circuit, the lithium-ion battery model tested with a 300A pulse current, an analog PID controller circuit with a 575V battery pack, and an analog PID controller circuit with a 3.7V lithium-ion battery cell that is still in need of tuning.